refactor(test) : e2e 테스트 분리 및 리팩토링
